Abstract out Raptor FileRewriter
