+++ presto-raptor/src/main/java/com/facebook/presto/raptor/storage/OrcFileInfo.java
@@ -1,0 +1,39 @@
+/*
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.facebook.presto.raptor.storage;
+
+import javax.annotation.concurrent.Immutable;
+
+@Immutable
+public class OrcFileInfo
+{
+    private final long rowCount;
+    private final long uncompressedSize;
+
+    public OrcFileInfo(long rowCount, long uncompressedSize)
+    {
+        this.rowCount = rowCount;
+        this.uncompressedSize = uncompressedSize;
+    }
+
+    public long getRowCount()
+    {
+        return rowCount;
+    }
+
+    public long getUncompressedSize()
+    {
+        return uncompressedSize;
+    }
+}
